pyproject.toml
[tool.poetry]
name = "delay-ease"
version = "0.2.0"
description = ""
authors = ["Ethan P"]
readme = "README.md"
packages = [{include = "delay_ease", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
requests = ">=2.32.3,<3.0.0"
openai = ">=1.68.2,<2.0.0"
browser-use = ">=0.7.0,<0.8.0"
python-dotenv = "^1.0.0"
typer = ">=0.12,<0.13"
click = ">=8.1,<8.2"
langchain-openai = "^0.3"


[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src"]


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
